backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
pydantic[email]==2.7.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30

# Encryption AES-256
cryptography==42.0.8

# LLM
groq==0.9.0
ollama==0.2.1

# Embeddings + Vector DB (works on Python 3.13, no Rust/C++ needed)
sentence-transformers==2.7.0
qdrant-client==1.9.1
numpy

# Document parsing
striprtf==0.0.26
python-docx==1.1.2
pypdf==4.2.0

# Scraping
httpx==0.27.0
beautifulsoup4==4.12.3

# Scheduler
apscheduler==3.10.4

# Utilities
tqdm==4.66.4
